rgw: civetweb/openssl: Load libssl.so and libcrypto.so by soname.
If building with radosgw, always look for openssl library (even when
building with nss).  Then, use objdump to fetch SONAME from the copies
of libssl and libcrypto that were found.  When building civetweb; pass
the library soname values in as the libraries to load with "dlopen".

This is a problem that went away for a bit, but came back with some
changes for
http://tracker.ceph.com/issues/16535

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1341775
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1258961

Fixes: http://tracker.ceph.com/issues/11239

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit 7caa0bd002110b62514da83a37a2a3deb841267a)
(cherry picked from commit 045551485415309ba9bad77e1aee28a0008881ca)

Resolves: rhbz#1258961