radosgw: Remove the newline from swift errors.
The current code emits a newline after swift errors, but fails
to account for it when it calculates 'content-length'.  This results in
some clients (go github.com/ncw/swift) producing complaints about the
unsolicited newline such as this,
	Unsolicited response received on idle HTTP channel starting with "\n"; err=<nil>

This patch removes the newline, but only after get_len() was called,
which happens at the end of the request while calculating content-length.
There are other parts of the swift wire protocol that depend on the
newline being emitted, but all that happens before get_len() is called.

Fixes: http://tracker.ceph.com/issues/18473
Signed-off-by: Marcus Watts <mwatts@redhat.com>