os/bluestore: allow latency to adpatively manage the size of the max bytes throttle
Uses the commit latency of the transactions to adapt the max bytes
throttle setting. The maximum number of bytes will be set to a power
of the square root of 2. When the latency is above the maximum latency
setting, it increases the size of this throttle by one step. When the
latency is below the minimum latency setting it decreases the size of
this throttle by one step.

When latency managed throttle is engaged, the max ops, deferred
max ops, and deferred max bytes throttles are not used.

Adds four new configuration options (the last 3 can be changed while
running):

  * bluestore_latency_managed_throttle     : whether to use the managed throttle
  * bluestore_throttle_latency_min         : minimum latency
  * bluestore_throttle_latency_max         : maximum latency
  * bluestore_throttle_latency_window_msec : duration of window in which
                                             average latencies are computed in
                                             milliseconds

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>