mds: track dentries that should be trimmed quickly in separate list
To trim a null dentry quickly, current method is moving it to bottom
of LRU list. The method depends on the cache size config and if there
is non-null dentry at the bottom of LRU list (MDCache::trim() keeps
trimming null dentries regardless of cache size config. It stops when
encountering a non-null dentry). MDS can move dentry of any type to
bottom of the LRU, so the method does not work as expected. Besides,
it's hard to extend the code to trim other type of dentry quickly.

The fix is use separate list to track dentries that should be trimmed
quickly. Both null dentries and non-auth stray dentries are added to
the list.

Signed-off-by: "Yan, Zheng" <zyan@redhat.com>