tests: rados: sleep before ceph tell osd.0 flush_pg_stats after restart
Even though we wait for HEALTH_OK after restarting the daemons, they are not
ready to respond to flush_pg_stats.

The reason why the osd is not ready for "tell" command after "ceph health"
shows that the cluster is "HEALTH_OK" is that the monitor fails to be notified
that the osd in question is not up in "heatbeat_interval". Because infernalis
does not have the osd_fast_fail_on_connection_refused support, the monitor
needs longer to detect that an osd is down, and osd_heartbeat_grace is used to
determine if an osd is down.

References: http://tracker.ceph.com/issues/16239
Signed-off-by: Nathan Cutler <ncutler@suse.com>
Signed-off-by: Kefu Chai <kchai@redhat.com>