mgr: Fix handle_command
We can't simply return failure if the command is not an mgr command
because in that case, we will never get to the python commands and their
lookup will fail. The patch fixes that by checking caps only for the
commands that can be found and returns 'not supported' only if the
handle_command function reaches its end.

Signed-off-by: Boris Ranto <branto@redhat.com>