os/bluestore/BlockDevice: fix racy aio_{wake,wait}() on IOContext
Thread 1 (_do_read)                 Thread 2 (_aio_thread)
queues aio
ioc->aio_wait()
  locks ioc->lock
  num_waiting++
                                    finishes aios
                                    ioc->aio_wake
                                      if (num_waiting)
                                        lock ioc->lock (blocks)
  num_running == 0, continue
  ioc->unlock
                                        ioc->lock taken
do_read destroys ioc
                                        use after free, may block forever

The last bit of the race may vary; the key thing is that thread 2 is
taking a lock that thread 1 can destroy; thread 2 doesn't have it pinned
in memory.

Fix this by simplifying the aio_wake, aio_wait.  Since it is mutually
exclusive with a callback completion, we can avoid calling it at all when
a callback in present, and focus on keeping it simple.

We avoid use-after-free by incrementing unwoken before decrementing
num_running, and preventing aio_wait() from returning until num_running is
zero *and* unwoken is zero.  Since the last thing that aio_wake() does to
touch ioc is to release the lock, and unwoken is decremented under the
lock, we can be sure the aio_wait() caller doesn't free the ioc before
io_thread is done with it.

Signed-off-by: Sage Weil <sage@redhat.com>