mds: match EMetaBlob::add_dir_context with MDCache::create_subtree_map
In EMetaBlob::add_dir_context(), the code that checks if a subtree
was journaled in the latest ESubtreeMap is different from the code
that creates ESubtreeMap. When creating ESubtreeMap, if a subtree's
dir_auth.first is equal to its parent subtree's dir_auth.first, the
subtree may not be journaled. EMetaBlob::add_dir_context() attempts
to get the result by checking CDir::STATE_{EXPORTBOUND,AUXSUBTREE}
and CDir::is_ambiguous_auth(). The difference makes the code hard
to understand and error prone.

Signed-off-by: "Yan, Zheng" <zyan@redhat.com>