osd: behave better when the snap reservation callback races with being canceled
Canceling a callback via the AsyncReserver unfortunately doesn't guarantee
the callback won't be executed. (If it's already been given to the Finisher
thread, there's not an available trail to follow to invalidate it -- and it
might already be executing concurrently anyway.) If we happen to cancel it
after the point of no return, we can't know when it's done -- and in the case
of shutdown, we really want that PGRef to go away.

Make this much, much safe by adding a separate lock to the ReservationCB which
protects its PGRef and "canceled" bool. When finishing, check if canceled, copy
the PGRef to a function-local member to prevent anything deleting it out from
under us, drop the lock and acquire the pg lock, then double-check if we're
canceled and otherwise do the operation.
This closes the race from logically concurrent to actually simultaneous threads,
and for shutdown we are okay with that.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>