os/bluestore: fix dirty_shard off-by-one
If we dirty an extent that abuts the end of a
shard, we should not dirty the next shard too.

This fixes a failed assertion in the unsharing code where
we are dirtying the range for a unshared extent and dirty_range
tries to dirty the next shard too, failing with an assertion that
it is not loaded.

Fixes: http://tracker.ceph.com/issues/20849
Signed-off-by: Sage Weil <sage@redhat.com>